+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ### Changed
 
 - Moved SettingsParser and setting to a new module
+- Renamed setting to get_setting
 
 ### Removed